requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
pydantic==2.5.0
python-dateutil==2.8.2
httpx==0.25.2
tenacity==8.2.3
python-multipart==0.0.6
selenium==4.15.2
fake-useragent==1.4.0
python-dotenv==1.0.0